Serialise player using Bitstream
Note that size() function returns something of type size_t which on
a 64 bit system is 8 bytes. This was causing me a major problem as
i thought it was a 32 bit integer. Hence all of my memory reading
was out of sync.

+++ GD4_Core_26/MemoryBitStream.hpp
@@ -64,7 +64,7 @@ public:
 	template< typename T >
 	void Write(const std::vector< T >& inVector)
 	{
-		uint32_t elementCount = inVector.size();
+		uint32_t elementCount = static_cast<uint32_t>(inVector.size());
 		Write(elementCount);
 		for (const T& element : inVector)
 		{
@@ -134,7 +134,7 @@ public:
 	template< typename T >
 	void Read(std::vector< T >& outVector)
 	{
-		size_t elementCount;
+		uint32_t elementCount;
 		Read(elementCount);
 		outVector.resize(elementCount);
 		for (T& element : outVector)
