+++ GD4_Core_26/player.hpp
@@ -3,8 +3,8 @@ class Player
 {
 public:
 	Player();
-	uint16_t GetHealth() const;
-	uint16_t GetAmmo() const;
+	uint32_t GetHealth() const;
+	uint32_t GetAmmo() const;
 	void Write(OutputMemoryStream& out_stream) const;
 	void Read(InputMemoryStream& in_stream);
 	void Write(OutputMemoryBitStream& out_stream) const;
@@ -14,8 +14,8 @@ public:
 
 
 private:
-	uint16_t health_;
-	uint16_t ammo_;
+	uint32_t health_;
+	uint32_t ammo_;
 	char name_[128];
 	Vector3 position_;
 	Quaternion rotation_;
